LPD-99383 Fold the FIPS audit sink into the emitter
FIPSAuditUtil was public in an exported package, so write took a severity and an arbitrary map from any OSGi module and put the result straight in the trail. That is the exposure c8987937 closed from the other direction when it nested caller fields under a reserved fields key so an event could not overwrite an envelope key; a public sink reopened it, because a caller that skips the emitter has no envelope at all and the record cannot be tied to the module or deployment that wrote it. Folding the sink in leaves emit as the only way to reach the trail, so every record is assembled by the code that owns the envelope.

The cost is that com.liferay.portal.kernel.security.fips is now the first exported portal-kernel package to import Log4J. Every one of the nine classes that touch it lives in com.liferay.portal.kernel.internal.log4j today, and Log4JUtil is the same shape as this class: a public facade in an exported package that delegates to internal.log4j and imports nothing from Log4J itself. Keeping the sink in an unexported package would have preserved that, but it only stops other bundles, since the class stays public on the portal class loader and portal-impl, portal-web and util-java could still call it. A single public entry point is the stronger guarantee and the easier one to state in a security policy, so it wins over the package boundary.

The severity is resolved once because _write needs it, but the provider and the timestamp keep their suppliers. MapBuilderCheck rejects a local that is only read inside the builder and points at a lambda instead, so hoisting them is not available; the provider is read twice as a result, which costs two copies of the JCE provider array per record and cannot make the name and version disagree while the list is fixed at boot.

Both test classes lose the FIPSAuditUtil static mock and take log4j as their seam instead. FIPSAuditEventEmitterUtilTest captures the ObjectMessage the logger receives, which also lets it assert the severity to level mapping that used to be checked separately, and its two level cases collapse into one helper. FIPSApplicationStateMachineUtilTest keeps every assertion it had: capturing the same record map one layer lower means only its setup changed, and it now exercises the real envelope on the way through.
